+++ app/Configuration.php
@@ -21,6 +21,7 @@ class Configuration extends Model
         'loglevel',
         'username',
         'password',
+        'twittertoken',
         'xmppdomain',
         'xmppdescription',
         'xmppwhitelist'