+++ database/migrations/20220722133156_add_twitter_token_to_configuration_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddTwitterTokenToConfigurationTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('configuration', function (Blueprint $table) {
+            $table->string('twittertoken')->nullable();
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('configuration', function (Blueprint $table) {
+            $table->dropColumn('twittertoken');
+        });
+    }
+}