+++ src/Movim/Bootstrap.php
@@ -312,8 +312,6 @@ class Bootstrap
     {
         if (\is_array($trace)) $trace = '';
 
-        echo 'An error occured, check syslog for more information'."\n";
-
         $error = $errstr . " in " . $errfile . ' (line ' . $errline . ")\n" . 'Trace' . "\n" . $trace;
 
         if (class_exists('Utils')) {