+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.21 (trunk)
  * Implement XEP-0461: Message Replies
  * Add PWA Push Notification support through the service worker
  * CSS fixes
+ * Upgrade Embed to v4.x and add Twitter integration
 
 v0.20
 ---------------------------